upcode

+++ app/Http/Controllers/BlogController.php
@@ -17,7 +17,7 @@ class BlogController extends Controller
     //select all
     public function index()
     {
-        $blogs = User::find(Auth::user()->id)->blogs;
+        $blogs = Blog::where('user_id',Auth::user()->id)->paginate(10);
         return View::make('blog.danhsach', compact('blogs'));
     }
 
@@ -107,17 +107,33 @@ class BlogController extends Controller
     //xoa nhieu
     public function destroyall(Request $request)
     {
-        $ids = $request->ids;
-        Blog::whereIn('id', $ids)->delete();
-        return redirect()->route('blog1.list');
+        if(Blog::where('user_id', Auth::user()->id)){
+            $ids = $request->ids;
+            Blog::whereIn('id', $ids)->delete();
+            return redirect()->route('blog1.list');
+        }else{
+            return view('404');
+        }
+    }
+    //thung rac
+    public function blog_trash()
+    {
+        $blogs_trash = Blog::onlyTrashed()->where('user_id',Auth::user()->id)->paginate(10);
+        return View::make('blog.blogs_trash', compact('blogs_trash'));
+    }
+    //khoi phuc
+    public function blog_untrash($id)
+    {
+        $blog = Blog::onlyTrashed()->where('user_id',Auth::user()->id)->find($id);
+        $blog->restore();
+        return redirect()->route('blog1.list')->with('ms', 'Khôi phục thành công');
     }
     //khoi phuc tin da xoa
     public function restore()
     {
-        Blog::onlyTrashed()->restore();
-        return redirect()->route('blog1.list');
+        Blog::onlyTrashed()->where('user_id',Auth::user()->id)->restore();
+        return redirect()->route('blog1.list')->with('ms', 'Khôi phục thành công');
     }
-
     public function search(Request $request)
     {
         $keywords = $request->keywords_submit;
@@ -125,7 +141,33 @@ class BlogController extends Controller
 
         return view('tim-kiem-blog',)->with('search_blog', $search_blog);
     }
-
+    public function search_ajax(Request $request)
+    {
+        if ($request->ajax()) {
+            $output = '';
+            $blogs=Blog::where('tieude','like','%'.$request->search.'%')
+                        ->orwhere('tennguoiviet',$request->search)->get();
+            foreach ($blogs as  $al) {
+                                $i=1;
+                                $output .= '<tr>
+                                <td><input type="checkbox" name="ids" class="checkBoxClass" value="{{$al->id}}"></td>
+                                <td>'.$al->id.'</td>
+                                <td>'.$al->tieude.'</td>
+                                <td>'.$al->tennguoiviet.'</td>
+                                <td>'.$al->noidung.'</td>
+                                <td><img src="{{ asset(anh_blog/'.$al->anh.') }}" style="width:90px; height: 80px;" alt=""></td>
+                                <td>'.'
+                                        <form action="/blog/xoa-blog/'.$al->id.'" method="post">
+                                        <input name="_method" type="hidden" value="DELETE">
+                                        <button type="submit" class="btn btn-xs btn-danger btn-flat show_confirm" data-toggle="tooltip" title="Delete"><i class="fa fa-trash" aria-hidden="true"></i></button>
+                                        </form>
+                                   '.'</td>
+                                   <td>'.'<a href="/blog/cap-nhat-blog/'.$al->id.'"><button class="btn btn-primary"><i class="fas fa-edit"></i></button></a>'.'</td>
+                                </tr>';
+                            }
+                        }
+                        return Response($output);
+                    }
     public function postComment($id, Request $request)
     {
         $comment = new BinhLuan;
